refactor: displayClassSequence 타입 변경

+++ TimFresh_Assignment/DataLayer/DTO/Request/AppDisplayClassInfoBySubDisplayClassInfoRequestDTO.swift
@@ -8,5 +8,5 @@
 import Foundation
 
 struct AppDisplayClassInfoBySubDisplayClassInfoRequestDTO {
-    let displayClassSequence: Int
+    let displayClassSequence: Int64
 }
